+++ podcasts/Episode/EpisodeLoadingController.swift
@@ -76,20 +76,9 @@ class EpisodeLoadingController: UIHostingController<AnyView> {
     }
 
     func loadEpisode() async -> Bool {
-        await withCheckedContinuation { continuation in
-            // If we're missing the podcast, then load that and the episode
-            if self.podcast == nil {
-                ServerPodcastManager.shared.addMissingPodcastAndEpisode(episodeUuid: episodeUuid, podcastUuid: podcastUuid)
-            }
-            // If we're missing just the episode then get that
-            else {
-                _ = ServerPodcastManager.shared.addMissingEpisode(episodeUuid: episodeUuid, podcastUuid: podcastUuid)
-            }
+        _ = try? await ServerPodcastManager.shared.addMissingPodcastAndEpisode(episodeUuid: episodeUuid, podcastUuid: podcastUuid)
 
-            // Verify they were added
-            let success = podcast != nil && episode != nil
-            continuation.resume(with: .success(success))
-        }
+        return podcast != nil && episode != nil
     }
 
     @MainActor func doneLoading() {