+++ podcasts/PlayerChapterCell.swift
@@ -156,8 +156,12 @@ class PlayerChapterCell: UITableViewCell {
     }
 
     @IBAction func linkTapped(_ sender: Any) {
-        guard let link = chapter?.url, let url = URL(string: link), let linkTapped = onLinkTapped else { return }
-        PlaybackManager.shared.trackChapterEvent(.chapterLinkClicked)
+        guard let chapter, let link = chapter.url, let url = URL(string: link), let linkTapped = onLinkTapped else { return }
+        PlaybackManager.shared.trackChapterEvent(.chapterLinkClicked, properties: [
+            "podcast_uuid": PlaybackManager.shared.currentPodcast?.uuid ?? "unknown",
+            "episode_uuid": PlaybackManager.shared.currentEpisode()?.uuid ?? "unknown",
+            "chapter_title": chapter.title
+        ])
         linkTapped(url)
     }